Install `fake-indexeddb` for testing IndexedDB
IndexedDB doesn't exist in Node environments, so we need something to
mock it for testing. We will need to be careful that we're not relying
on the mock when implementing and testing features intended to work in
server-side rendering situations.

One of `fake-indexeddb`'s dependencies (`base64-arraybuffer-es6`)
requires `core-js-bundle` as a peer dependency, so we've installed it.
Probably `fake-indexeddb` should be installing it for us, really.