Create a Database wrapper around `idb`'s IndexedDB wrapper
For now it only opens and closes. More methods incoming.

`#close()` should really wait for the IndexedDB connection to close, but
`fake-indexeddb` doesn't implement the `close` event that would be
needed to do that, so we wouldn't be able to test it properly. In
practice we're unlikely to be closing a database and reopening before
any transactions complete (which is what blocks closing), as
transactions complete themselves when not given things to do.

The fact that `fake-indexeddb` doesn't fully match the API expected in a
browser means we should make sure to run some tests around this in a
browser environment, or find a more trustworthy mock implementation for
Node environment testing.

For now, the wrapper exposes internal `idb` types in the callbacks. We
probably don't want to have to expose the internals of `idb` to end
users, so we'll want to wrap them.