Add a `transaction` method to `Database`
This is for those times where a user wants to do more complex operations
than a single put or get. Transactions close themselves if they stop
being interacted with, so we encourage transactions to be self contained
by wrapping them in a callback that prefetches the store.

This exposes the internal object store to users, which we probably want
to wrap, so they don't need to know anything about `idb`.

The fact that `fake-indexeddb` doesn't support `close` events means we
do some hacky stuff to test that `#close()` waits for transactions to
finish before closing the connection.