Create an `expect` wrapper to simplify awaiting promises
When using `expect` to test the result of a promise, you need to pass it
an uncomplete promise. Sometimes promise implementations throw or exit
immediately in some cases, so it's common to wrap them in immediately
invoked function expressions that return a promise that explicitly
awaits the promise you want to test. `expectPromise` makes this less
arduous to do, by providing a wrapper that expects a callback to the
promise, rather than a full async IIFE.