Create a `NamedSchema` interface to allow `Database` to type name
Given opening a database means permanently creating an IndexedDB that
persists across sessions, we need to be sure we're always using the same
name for that database. Using `NamedSchema` allows us to infer the
possible values for the name, and prevent calls with a name that doesn't
match those values.