Stop `openDB` from hanging if there are issues with its callbacks
Under the hood, IndexedDB relies on event emitters, which are loosly
wrapped in a promise by `idb`. `idb` doesn't link the `upgrade` or
`blocked` events into that promise, however, but errors in those
callbacks can prevent the promise from ever settling. We wrap the call
to `openDB` in our own promise, and hook into its event emitter to catch
the error cases properly.

We don't hook the `blocking` callback into the promise, as it will only
be called on an already open database, so this promise should be long
resolved by that point.