Wrap the arguments of the upgrade callback and handle promises
We want to be able to use a similar interface during upgrades as we do
normally. While a database is upgrading is the only time stores and
indexed can be created.

While we add support for promises to `upgrade` and `blocked`, we don't
support them in `blocking` as that call needs to be synchronous or it
will continue working after the blocked database has given up and
thrown, removing most of the benefit of being able to handle blocking
at all.