Hardcode the path to `idb`'s index file
Something weird is going on with `idb`. VSCode has trouble finding it
sometimes, and TypeDoc (spoilers) is completely unable to find it. By
setting the path in the `tsconfig.json`, we tell TypeScript how to find
the module. This shouldn't be necessary, but I have been unable to find
why `idb` is special and causing problems.

Fixing the imports also fixes ESLint complaining about a necessary type
assertion that it thought was unnecessary.