Make Dependabot update less frequently
While always being up-to-date with cutting edge releases of packages is
nice, it can also result in a lot of updates. Most don't require manual
reviews, but because they're automerged all the time, it can make
keeping up with what's changing quite difficult.

This is a library that will be published preiodically. It's not
necessary to be absolutely up-to-date all of the time, so we can make
the schedule daily (or weekly for examples where it matters even less).