Create a database provider
This uses a `DatabaseContext` to provide the database to its children.
The component wait for the `Database.open` promise to settle, and
updates the context with the result.

To test the context, we now have a test context consumer.

To test the error cases of React components, it was necessary to create
an error boundary.