Create dynamic component wrappers to interface with a database
This adds the concept of a `DynamicPageComponent`, a component that
and fetches and updates a single value of a `Database` store. The
component is wrapped in a `WrappedPageComponent` to hook map its props
to the database.

To differentiate "dynamic" from "static" components, `PageComponent` has
been renamed to `StaticPageComponent`.

`WrappedPageComponent` relies on a wrapping `DatabaseContext` to access
the database. It's disabled while waiting for database operations to
settle, to prevent user input from overwriting stored data. It supports
a `defaultValue` to display if there is no stored data. It fetches from
the database as soon as it has a database to fetch from, but after that,
never updates unless it is remounted.

It does not currently persist its data at all. That's still to come, as
is fetching updated values from the store. The `Orchestrator` is also
not set up to do the `DatabaseContext` wrapping, yet.

This is a bit of a large commit, sorry.