Make `Page` update the database on submission
We require users to provide the submit button component, so they can
hook up any routers, or styling they need. We might want to improve this
with a default in the future.

If the current value matches the empty value, we consider that to be a
deletion.

This also renames "key" to "slug" for steps, as that's what they really
represent.