Add support for objects as values in stores
To do this, we add some new types, and pull `Store` out into its own
wrapper, along with some other `idb` types. This will help us control
the API of the store better.

There's a bunch of casting to deal with not having specific types for
the different levels. We support a maximum depth of 2 levels in value
objects.